requirements.txt
waveshare-epaper # uses rpo.gpio - only installable on raspberry pi
pillow==9.5.0
requests